move HTML generators to DocGen class so file manager can focus on managing files
